refactor: move response to dtos
